requirements.txt
boto3==1.34.131
botocore==1.34.131

# LangChain modern stack
langchain==0.2.16
langchain-core==0.2.38
langchain-aws==0.2.33
langsmith==0.1.108

# Otros útiles para tu proyecto
pillow==9.5.0
numpy==1.26.4
python-dotenv==1.0.0
pytest==7.4.2
langchain-community